python/requirements.txt
pytest-repeat
numpy==1.19.0